Add CancellationToken, ownership checks, and integration tests
- Add CancellationToken support to all async methods and Dapper calls
- Enforce parent-child ownership in room, parking, contact, and outdoor services
- Refactor RoomDto construction into RoomDtoBuilder for parallel loading
- Change ConditionRating to decimal? for fractional ratings
- Make listing reference number generation concurrency-safe with SQL locks
- Enhance ExceptionHandlingMiddleware for SQL constraint violations (409/400)
- Centralize DI registrations in DependencyInjection.cs
- Add integration tests with test DB and schema via DatabaseFixture
- Embed Schema.sql for consistent test DB setup
- Update .csproj/.slnx for test files and resources

+++ Application/DTOs/RoomDtos.cs
@@ -29,7 +29,7 @@ public record RoomDto(
 );
 
 public record UpsertRoomConditionRequest(
-    int? ConditionRating,
+    decimal? ConditionRating,
     string? Notes,
     int ConditionCategoryId
 );
@@ -37,7 +37,7 @@ public record UpsertRoomConditionRequest(
 public record RoomConditionDto(
     int Id,
     int ListingRoomId,
-    int? ConditionRating,
+    decimal? ConditionRating,
     string? Notes,
     int ConditionCategoryId
 );
