+++ Domain/Models/Condition.cs
@@ -4,7 +4,7 @@ public class Condition
 {
     public int Id { get; set; }
     public int ListingRoomId { get; set; }
-    public int? ConditionRating { get; set; }
+    public decimal? ConditionRating { get; set; }
     public string? Notes { get; set; }
     public int ConditionCategoryId { get; set; }
 }
